eval: fold weight_target into checkpoint_eval
